Refactor cross-section classes: replace RHSCFCornerCrossSection with QuarterCircularSpandrelCrossSection and update related tests and fixtures
